requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==24.1